+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+* Unreachable ad media (Velocity error `2013`) is now reported to MAX as `noFill` so the waterfall moves on.
+
 ## [0.10.0.0] - 2026-09-07
 
 ### Added